Entity Interface Implemented
Along with the implementation of the Entity Interface are some minor tweaks to some of the code in the controller and unit classes.  RTS Controller now has a debug function that lets you damage your selected units and structure using Numpad 1.   When an Entities' health reaches or falls below 0, it will be destroyed.