Fixed bugs, added particles
I fixed some strange bugs that are showing up with the meshes by adding #include "Engine/StaticMesh.h" to everything that was experiencing an issue.

I disabled the InitializeEntity in the Rifleman, Rocketeer, and Engineer as they may have been conflicting with the Barracks code.

I added multiple particle effects to the asset folder and have implemented two of them, one in the rifleman and the other in the rocketeer.

I have put in first pass code to spawn units from the barracks, however they can't find the navmesh and you can only spawn one before the game crashes for some reason.