Update: Tues 7:41pm
Created the Humvee/Scout units (which was forgotten before)

Created and Implemented models for the Humvee, Mobile Artillery, and the Mobile Outpost.

Updated some of the code from the MBT, AvBT, and the Gattling Tank to allow the rotation of the guns towards their target and some adjustments to the targeting systems.