Update: Wed, 7:53pm
Added targeting system for Harvester that allows it to harvest a specific resource node and to return to a refinery by player commands.

NOTE:  Need to fix the hitboxes for the resource nodes.  It works, but the node you click isn't the node that the harvester goes to.  Suspected that it's a raycast issue on the hitbox of the resource node.