Update: Mon, 1:43pm
Implemented optional destructibility for entities.  It's not possible to make entities indestructible / unkillable.

Implemented a crash fix for harvesters that don't have a refinery to return to.

NOTE: Found bug in harvester but didn't get the chance to fix it in time.  Don't inherit refinery / harvester pairs for the time being.  There seems to be an issue that's causing the harvester to be inherited but not the refinery.