BUG TRACKER 2
Lowered the color brightness (not the emission) on the resource materials.  It still emits like it did before, but just not as bright of a color.

Fixed the outpost crashing issue.  It seems our selectedcharacters array is holding onto null reference pointers.  This is something we really need to look into in the future.  We dont want to be allocating memory to a list of empty objects.